// src/enums/PostType.ts
export enum PostType {
  TEXT = "TEXT", // Sadece metin
  IMAGE = "IMAGE", // Resim
  VIDEO = "VIDEO", // Video
  GALLERY = "GALLERY", // Galeri
  LINK = "LINK", // Bağlantı
  EVENT = "EVENT", // Etkinlik
  ANNOUNCEMENT = "ANNOUNCEMENT", // Duyuru
  NEWS = "NEWS", // Haber
  ACHIEVEMENT = "ACHIEVEMENT", // Başarı
  CELEBRATION = "CELEBRATION", // Kutlama
  POLL = "POLL", // Anket
  QUOTE = "QUOTE", // Alıntı
  TESTIMONIAL = "TESTIMONIAL", // Referans
  BEHIND_SCENES = "BEHIND_SCENES", // Kuliste
  LIVE_STREAM = "LIVE_STREAM", // Canlı yayın
}