// src/enums/PostStatus.ts
export enum PostStatus {
  DRAFT = "DRAFT", // Taslak
  SCHEDULED = "SCHEDULED", // Zamanlanmış
  PUBLISHED = "PUBLISHED", // Yayınlanmış
  ARCHIVED = "ARCHIVED", // Arşivlenmiş
  DELETED = "DELETED", // Silinmiş
  MODERATION = "MODERATION", // Moderasyonda
  REJECTED = "REJECTED", // Reddedilmiş
  EXPIRED = "EXPIRED", // Süresi dolmuş
}